fixed broken existing tests and linted
