+++ tethysapp/app_store/begin_install.py
@@ -1,5 +1,4 @@
 import os
-import time
 import importlib
 import subprocess
 import tethysapp
@@ -13,7 +12,7 @@ from subprocess import call
 from .helpers import check_all_present, logger, send_notification
 from .resource_helpers import get_resource, get_app_instance_from_path
 from .proxy_app_handlers import create_proxy_app, list_proxy_apps
-from .mamba_helpers import mamba_uninstall, mamba_download, mamba_install
+from .mamba_helpers import mamba_download, mamba_install
 from tethys_apps.base.workspace import TethysWorkspace
 
 
@@ -137,10 +136,9 @@ def begin_install(installData, channel_layer, app_workspace):
     send_notification(f"Installing Version: {installData['version']}", channel_layer)
 
     try:
-        
         if resource['app_type'] == "tethysapp":
             successful_install = mamba_install(resource, installData['channel'], installData['label'],
-                                            installData["version"], channel_layer)
+                                               installData["version"], channel_layer)
             if not successful_install:
                 raise Exception("Mamba install script failed to install application.")
             detect_app_dependencies(resource['name'], channel_layer)
@@ -151,9 +149,9 @@ def begin_install(installData, channel_layer, app_workspace):
                 message = "Proxy App is already installed with this name"
                 send_notification(message, channel_layer)
                 raise Exception(message)
-            
+
             successful_install = mamba_download(resource, installData['channel'], installData['label'],
-                                            installData["version"], channel_layer)
+                                                installData["version"], channel_layer)
 
             site_packages = os.path.join(os.path.dirname(subprocess.__file__), "site-packages")
             proxy_package = [package for package in os.listdir(site_packages) if resource["name"] in package][0]
@@ -171,7 +169,7 @@ def begin_install(installData, channel_layer, app_workspace):
                 "jsHelperFunction": "proxyAppInstallComplete",
                 "helper": "addModalHelper"
             }
-            send_notification(get_data_json, channel_layer) 
+            send_notification(get_data_json, channel_layer)
     except Exception as e:
         logger.error(e)
         send_notification("Application installation failed. Check logs for more details.", channel_layer)
