+++ tethysapp/app_store/tests/unit_tests/test_helpers.py
@@ -4,7 +4,8 @@ from pathlib import Path
 from unittest.mock import MagicMock, call
 from tethysapp.app_store.helpers import (parse_setup_file, get_conda_stores, check_all_present, run_process,
                                          send_notification, apply_template, get_github_install_metadata,
-                                         get_override_key, get_color_label_dict, get_setup_path)
+                                         get_override_key, get_color_label_dict, get_setup_path, restart_server,
+                                         CACHE_KEY, clear_github_cache_list)
 
 
 def test_get_override_key(mocker):
@@ -278,3 +279,152 @@ def test_get_setup_path_missing_file(tmp_path):
         get_setup_path(str(tmp_path))
 
     assert e.value.args[0] == 'Unable to find a project file for application'
+
+
+def test_restart_server_dev_server(mocker, caplog, tmp_path):
+    app_files = tmp_path / "tethysapp" / "app_store"
+    app_files.mkdir(parents=True)
+    function_file = app_files / "fake_file.py"
+    mock_ws = mocker.patch('tethysapp.app_store.helpers.send_notification')
+    mock_run_process = mocker.patch('tethysapp.app_store.helpers.run_process')
+    mocker.patch('tethysapp.app_store.helpers.get_manage_path', return_value="manage_path")
+    mocker.patch('tethysapp.app_store.helpers.sys.argv', ["manage_path", "runserver"])
+    mocker.patch('tethysapp.app_store.helpers.os.path.realpath', return_value=function_file)
+    data = {"name": "test_app", "restart_type": "install"}
+    mock_channel = MagicMock()
+    mock_workspace = MagicMock(path=str(tmp_path))
+
+    restart_server(data, mock_channel, mock_workspace)
+
+    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
+    mock_ws.assert_called_with(f"Running Syncstores for app: {data['name']}", mock_channel)
+    mock_run_process.assert_called_with(['python', "manage_path", 'syncstores', data["name"], '-f'])
+    assert "Dev Mode. Attempting to restart by changing file" in caplog.messages
+    model_py = app_files / "model.py"
+    assert model_py.read_text() == f'print("{data["name"]} installed in dev mode")\n'
+
+
+def test_restart_server_dev_server_install_scaffold_running(mocker, caplog, tmp_path):
+    app_files = tmp_path / "tethysapp" / "app_store"
+    app_files.mkdir(parents=True)
+    function_file = app_files / "fake_file.py"
+    mock_ws = mocker.patch('tethysapp.app_store.helpers.send_notification')
+    mock_run_process = mocker.patch('tethysapp.app_store.helpers.run_process')
+    mocker.patch('tethysapp.app_store.helpers.get_manage_path', return_value="manage_path")
+    mocker.patch('tethysapp.app_store.helpers.sys.argv', ["manage_path", "runserver"])
+    mocker.patch('tethysapp.app_store.helpers.os.path.realpath', return_value=function_file)
+    data = {"name": "test_app", "restart_type": "install"}
+    mock_channel = MagicMock()
+    mock_workspace = MagicMock(path=str(tmp_path))
+    install_status_dir = tmp_path / "install_status"
+    install_status_dir.mkdir(parents=True)
+    installRunning = install_status_dir / "installRunning"
+    installRunning.touch()
+    assert installRunning.is_file()
+
+    scaffoldRunning = install_status_dir / "scaffoldRunning"
+    scaffoldRunning.touch()
+    assert scaffoldRunning.is_file()
+
+    restart_server(data, mock_channel, mock_workspace)
+
+    assert not installRunning.is_file()
+    assert not scaffoldRunning.is_file()
+    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
+    mock_ws.assert_called_with(f"Running Syncstores for app: {data['name']}", mock_channel)
+    mock_run_process.assert_called_with(['python', "manage_path", 'syncstores', data["name"], '-f'])
+    assert "Dev Mode. Attempting to restart by changing file" in caplog.messages
+    model_py = app_files / "model.py"
+    assert model_py.read_text() == f'print("{data["name"]} installed in dev mode")\n'
+
+
+def test_restart_server_prod_server_run_collect_all(mocker, caplog, tmp_path):
+    mock_ws = mocker.patch('tethysapp.app_store.helpers.send_notification')
+    mock_run_process = mocker.patch('tethysapp.app_store.helpers.run_process')
+    mock_subprocess = mocker.patch('tethysapp.app_store.helpers.subprocess')
+    mocker.patch('tethysapp.app_store.helpers.app.get_custom_setting', return_value="custom_setting")
+    mocker.patch('tethysapp.app_store.helpers.get_manage_path', return_value="manage_path")
+    mock_os_system = mocker.patch('tethysapp.app_store.helpers.os.system')
+    data = {"name": "test_app", "restart_type": "install"}
+    mock_channel = MagicMock()
+    mock_workspace = MagicMock(path=str(tmp_path))
+
+    restart_server(data, mock_channel, mock_workspace)
+
+    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
+    mock_ws.assert_has_calls([
+        call(f"Running Syncstores for app: {data['name']}", mock_channel),
+        call(f"Running Tethys Collectall for app: {data['name']}", mock_channel),
+        call("Server Restarting . . .", mock_channel)
+    ])
+    mock_run_process.assert_has_calls([
+        call(['python', "manage_path", 'syncstores', data["name"], '-f']),
+        call(['python', "manage_path", 'pre_collectstatic']),
+        call(['python', "manage_path", 'collectstatic', '--noinput']),
+        call(['python', "manage_path", 'collectworkspaces', '--force'])
+    ])
+    mock_subprocess.run.assert_called_with(['sudo', '-h'], check=True)
+    mock_os_system.assert_called_with('echo custom_setting|sudo -S supervisorctl restart all')
+    assert "Running Tethys Collectall" in caplog.messages
+
+
+def test_restart_server_prod_server_docker(mocker, caplog, tmp_path):
+    mock_ws = mocker.patch('tethysapp.app_store.helpers.send_notification')
+    mock_run_process = mocker.patch('tethysapp.app_store.helpers.run_process')
+    mocker.patch('tethysapp.app_store.helpers.subprocess.run', side_effect=[Exception("No sudo")])
+    mocker.patch('tethysapp.app_store.helpers.get_manage_path', return_value="manage_path")
+    mocker.patch('tethysapp.app_store.helpers.os.path.isdir', return_value=True)
+    restart = tmp_path / "restart"
+    mocker.patch('tethysapp.app_store.helpers.Path', return_value=restart)
+    data = {"name": "test_app", "restart_type": "install"}
+    mock_channel = MagicMock()
+    mock_workspace = MagicMock(path=str(tmp_path))
+
+    restart_server(data, mock_channel, mock_workspace, run_collect_all=False)
+
+    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
+    mock_ws.assert_has_calls([
+        call(f"Running Syncstores for app: {data['name']}", mock_channel),
+        call("Server Restarting . . .", mock_channel)
+    ])
+    mock_run_process.assert_has_calls([
+        call(['python', "manage_path", 'syncstores', data["name"], '-f'])
+    ])
+    assert "No sudo" in caplog.messages
+    assert "No SUDO. Docker container implied. Restarting without SUDO" in caplog.messages
+    assert "Restart Directory found. Creating restart file." in caplog.messages
+    assert restart.is_file()
+
+
+def test_restart_server_prod_server_docker_retry_no_sudo(mocker, caplog, tmp_path):
+    mock_ws = mocker.patch('tethysapp.app_store.helpers.send_notification')
+    mock_run_process = mocker.patch('tethysapp.app_store.helpers.run_process')
+    mocker.patch('tethysapp.app_store.helpers.subprocess.run', side_effect=[Exception("No sudo")])
+    mocker.patch('tethysapp.app_store.helpers.get_manage_path', return_value="manage_path")
+    mocker.patch('tethysapp.app_store.helpers.os.path.isdir', return_value=False)
+    mock_os_system = mocker.patch('tethysapp.app_store.helpers.os.system')
+    data = {"name": "test_app", "restart_type": "install"}
+    mock_channel = MagicMock()
+    mock_workspace = MagicMock(path=str(tmp_path))
+
+    restart_server(data, mock_channel, mock_workspace, run_collect_all=False)
+
+    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
+    mock_ws.assert_has_calls([
+        call(f"Running Syncstores for app: {data['name']}", mock_channel),
+        call("Server Restarting . . .", mock_channel)
+    ])
+    mock_run_process.assert_has_calls([
+        call(['python', "manage_path", 'syncstores', data["name"], '-f'])
+    ])
+    assert "No sudo" in caplog.messages
+    assert "No SUDO. Docker container implied. Restarting without SUDO" in caplog.messages
+    mock_os_system.assert_called_with('supervisorctl restart all')
+
+
+def test_clear_github_cache_list(mocker):
+    mock_cache = mocker.patch('tethysapp.app_store.helpers.cache')
+
+    clear_github_cache_list()
+
+    mock_cache.delete.assert_called_with(CACHE_KEY)
