+++ tethysapp/app_store/tests/unit_tests/test_installation_handlers.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock, call
 from argparse import Namespace
 from django.core.exceptions import ObjectDoesNotExist
 from tethys_sdk.app_settings import CustomSetting
-from tethysapp.app_store.installation_handlers import (get_service_options, restart_server, continueAfterInstall,
+from tethysapp.app_store.installation_handlers import (get_service_options, continueAfterInstall,
                                                        set_custom_settings, process_settings, configure_services,
                                                        getServiceList)
 
@@ -20,147 +20,6 @@ def test_get_service_options(mocker):
     assert services == expected_services
     expected_args = Namespace(spatial=True)
     mock_services_list_command.asserrt_called_with(expected_args)
-
-
-def test_restart_server_dev_server(mocker, caplog, tmp_path):
-    app_files = tmp_path / "tethysapp" / "app_store"
-    app_files.mkdir(parents=True)
-    function_file = app_files / "fake_file.py"
-    mock_ws = mocker.patch('tethysapp.app_store.installation_handlers.send_notification')
-    mock_run_process = mocker.patch('tethysapp.app_store.installation_handlers.run_process')
-    mocker.patch('tethysapp.app_store.installation_handlers.get_manage_path', return_value="manage_path")
-    mocker.patch('tethysapp.app_store.installation_handlers.sys.argv', ["manage_path", "runserver"])
-    mocker.patch('tethysapp.app_store.installation_handlers.os.path.realpath', return_value=function_file)
-    data = {"name": "test_app", "restart_type": "install"}
-    mock_channel = MagicMock()
-    mock_workspace = MagicMock(path=str(tmp_path))
-
-    restart_server(data, mock_channel, mock_workspace)
-
-    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
-    mock_ws.assert_called_with(f"Running Syncstores for app: {data['name']}", mock_channel)
-    mock_run_process.assert_called_with(['python', "manage_path", 'syncstores', data["name"], '-f'])
-    assert "Dev Mode. Attempting to restart by changing file" in caplog.messages
-    model_py = app_files / "model.py"
-    assert model_py.read_text() == f'print("{data["name"]} installed in dev mode")\n'
-
-
-def test_restart_server_dev_server_install_scaffold_running(mocker, caplog, tmp_path):
-    app_files = tmp_path / "tethysapp" / "app_store"
-    app_files.mkdir(parents=True)
-    function_file = app_files / "fake_file.py"
-    mock_ws = mocker.patch('tethysapp.app_store.installation_handlers.send_notification')
-    mock_run_process = mocker.patch('tethysapp.app_store.installation_handlers.run_process')
-    mocker.patch('tethysapp.app_store.installation_handlers.get_manage_path', return_value="manage_path")
-    mocker.patch('tethysapp.app_store.installation_handlers.sys.argv', ["manage_path", "runserver"])
-    mocker.patch('tethysapp.app_store.installation_handlers.os.path.realpath', return_value=function_file)
-    data = {"name": "test_app", "restart_type": "install"}
-    mock_channel = MagicMock()
-    mock_workspace = MagicMock(path=str(tmp_path))
-    install_status_dir = tmp_path / "install_status"
-    install_status_dir.mkdir(parents=True)
-    installRunning = install_status_dir / "installRunning"
-    installRunning.touch()
-    assert installRunning.is_file()
-
-    scaffoldRunning = install_status_dir / "scaffoldRunning"
-    scaffoldRunning.touch()
-    assert scaffoldRunning.is_file()
-
-    restart_server(data, mock_channel, mock_workspace)
-
-    assert not installRunning.is_file()
-    assert not scaffoldRunning.is_file()
-    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
-    mock_ws.assert_called_with(f"Running Syncstores for app: {data['name']}", mock_channel)
-    mock_run_process.assert_called_with(['python', "manage_path", 'syncstores', data["name"], '-f'])
-    assert "Dev Mode. Attempting to restart by changing file" in caplog.messages
-    model_py = app_files / "model.py"
-    assert model_py.read_text() == f'print("{data["name"]} installed in dev mode")\n'
-
-
-def test_restart_server_prod_server_run_collect_all(mocker, caplog, tmp_path):
-    mock_ws = mocker.patch('tethysapp.app_store.installation_handlers.send_notification')
-    mock_run_process = mocker.patch('tethysapp.app_store.installation_handlers.run_process')
-    mock_subprocess = mocker.patch('tethysapp.app_store.installation_handlers.subprocess')
-    mocker.patch('tethysapp.app_store.installation_handlers.app.get_custom_setting', return_value="custom_setting")
-    mocker.patch('tethysapp.app_store.installation_handlers.get_manage_path', return_value="manage_path")
-    mock_os_system = mocker.patch('tethysapp.app_store.installation_handlers.os.system')
-    data = {"name": "test_app", "restart_type": "install"}
-    mock_channel = MagicMock()
-    mock_workspace = MagicMock(path=str(tmp_path))
-
-    restart_server(data, mock_channel, mock_workspace)
-
-    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
-    mock_ws.assert_has_calls([
-        call(f"Running Syncstores for app: {data['name']}", mock_channel),
-        call(f"Running Tethys Collectall for app: {data['name']}", mock_channel),
-        call("Server Restarting . . .", mock_channel)
-    ])
-    mock_run_process.assert_has_calls([
-        call(['python', "manage_path", 'syncstores', data["name"], '-f']),
-        call(['python', "manage_path", 'pre_collectstatic']),
-        call(['python', "manage_path", 'collectstatic', '--noinput']),
-        call(['python', "manage_path", 'collectworkspaces', '--force'])
-    ])
-    mock_subprocess.run.assert_called_with(['sudo', '-h'], check=True)
-    mock_os_system.assert_called_with('echo custom_setting|sudo -S supervisorctl restart all')
-    assert "Running Tethys Collectall" in caplog.messages
-
-
-def test_restart_server_prod_server_docker(mocker, caplog, tmp_path):
-    mock_ws = mocker.patch('tethysapp.app_store.installation_handlers.send_notification')
-    mock_run_process = mocker.patch('tethysapp.app_store.installation_handlers.run_process')
-    mocker.patch('tethysapp.app_store.installation_handlers.subprocess.run', side_effect=[Exception("No sudo")])
-    mocker.patch('tethysapp.app_store.installation_handlers.get_manage_path', return_value="manage_path")
-    mocker.patch('tethysapp.app_store.installation_handlers.os.path.isdir', return_value=True)
-    restart = tmp_path / "restart"
-    mocker.patch('tethysapp.app_store.installation_handlers.Path', return_value=restart)
-    data = {"name": "test_app", "restart_type": "install"}
-    mock_channel = MagicMock()
-    mock_workspace = MagicMock(path=str(tmp_path))
-
-    restart_server(data, mock_channel, mock_workspace, run_collect_all=False)
-
-    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
-    mock_ws.assert_has_calls([
-        call(f"Running Syncstores for app: {data['name']}", mock_channel),
-        call("Server Restarting . . .", mock_channel)
-    ])
-    mock_run_process.assert_has_calls([
-        call(['python', "manage_path", 'syncstores', data["name"], '-f'])
-    ])
-    assert "No sudo" in caplog.messages
-    assert "No SUDO. Docker container implied. Restarting without SUDO" in caplog.messages
-    assert "Restart Directory found. Creating restart file." in caplog.messages
-    assert restart.is_file()
-
-
-def test_restart_server_prod_server_docker_retry_no_sudo(mocker, caplog, tmp_path):
-    mock_ws = mocker.patch('tethysapp.app_store.installation_handlers.send_notification')
-    mock_run_process = mocker.patch('tethysapp.app_store.installation_handlers.run_process')
-    mocker.patch('tethysapp.app_store.installation_handlers.subprocess.run', side_effect=[Exception("No sudo")])
-    mocker.patch('tethysapp.app_store.installation_handlers.get_manage_path', return_value="manage_path")
-    mocker.patch('tethysapp.app_store.installation_handlers.os.path.isdir', return_value=False)
-    mock_os_system = mocker.patch('tethysapp.app_store.installation_handlers.os.system')
-    data = {"name": "test_app", "restart_type": "install"}
-    mock_channel = MagicMock()
-    mock_workspace = MagicMock(path=str(tmp_path))
-
-    restart_server(data, mock_channel, mock_workspace, run_collect_all=False)
-
-    assert f"Running Syncstores for app: {data['name']}" in caplog.messages
-    mock_ws.assert_has_calls([
-        call(f"Running Syncstores for app: {data['name']}", mock_channel),
-        call("Server Restarting . . .", mock_channel)
-    ])
-    mock_run_process.assert_has_calls([
-        call(['python', "manage_path", 'syncstores', data["name"], '-f'])
-    ])
-    assert "No sudo" in caplog.messages
-    assert "No SUDO. Docker container implied. Restarting without SUDO" in caplog.messages
-    mock_os_system.assert_called_with('supervisorctl restart all')
 
 
 def test_continueAfterInstall(mocker):
