+++ tethysapp/app_store/tests/unit_tests/test_notifications.py
@@ -34,19 +34,15 @@ async def test_check_user_permissions_dne(mocker):
 
 
 @pytest.mark.asyncio
-async def test_notificationsConsumer_connect_disconnect(mocker, caplog):
-    mock_duser = mocker.patch('tethysapp.app_store.notifications.User')
-    mock_duser.objects.get().has_perm.return_value = True
-    mock_user = MagicMock(id=1)
+async def test_notificationsConsumer_connect_disconnect(caplog):
     consumer = notificationsConsumer
+    consumer._authorized = True
     consumer.channel_layer_alias = "testlayer"
     channel_layers_setting = {
         "testlayer": {"BACKEND": "channels.layers.InMemoryChannelLayer"}
     }
     with override_settings(CHANNEL_LAYERS=channel_layers_setting):
-        consumer = notificationsConsumer.as_asgi()
-        communicator = WebsocketCommunicator(notificationsConsumer.as_asgi(), "GET", "install/notifications")
-        communicator.scope["user"] = mock_user
+        communicator = WebsocketCommunicator(consumer.as_asgi(), "GET", "install/notifications")
         connected, _ = await communicator.connect()
         assert connected
         channel_layer = get_channel_layer("testlayer")
@@ -58,28 +54,26 @@ async def test_notificationsConsumer_connect_disconnect(mocker, caplog):
 
 
 @pytest.mark.asyncio
-async def test_notificationsConsumer_install_notifications(mocker, caplog):
-    mock_duser = mocker.patch('tethysapp.app_store.notifications.User')
-    mock_duser.objects.get().has_perm.return_value = True
-    mock_user = MagicMock(id=1)
+async def test_notificationsConsumer_install_notifications(caplog):
     consumer = notificationsConsumer
+    consumer._authorized = True
     consumer.channel_layer_alias = "testlayer"
     channel_layers_setting = {
         "testlayer": {"BACKEND": "channels.layers.InMemoryChannelLayer"}
     }
     with override_settings(CHANNEL_LAYERS=channel_layers_setting):
-        consumer = notificationsConsumer.as_asgi()
-        communicator = WebsocketCommunicator(notificationsConsumer.as_asgi(), "GET", "install/notifications")
-        communicator.scope["user"] = mock_user
+        communicator = WebsocketCommunicator(consumer.as_asgi(), "GET", "install/notifications")
         connected, _ = await communicator.connect()
         assert connected
 
         channel_layer = get_channel_layer("testlayer")
         channel_name = list(channel_layer.channels.keys())[0]
+        assert "notifications" in channel_layer.groups
 
         message = "Sending a message"
         await channel_layer.group_send("notifications", {"type": "install_notifications", "message": message})
-        await communicator.receive_from()
+        response = await communicator.receive_from()
+        assert response == json.dumps({'message': message})
 
         await communicator.disconnect()
         assert f"Added {channel_name} channel to notifications" in caplog.messages
@@ -89,23 +83,19 @@ async def test_notificationsConsumer_install_notifications(mocker, caplog):
 
 @pytest.mark.asyncio
 async def test_notificationsConsumer_receive_begin_install(mocker, caplog):
-    mock_duser = mocker.patch('tethysapp.app_store.notifications.User')
-    mock_duser.objects.get().has_perm.return_value = True
-    mock_user = MagicMock(id=1)
     mock_begin_install = mocker.patch('tethysapp.app_store.notifications.begin_install')
     mock_workspace = MagicMock()
     mock_get_workspace = AsyncMock(return_value=mock_workspace)
     mocker.patch('tethysapp.app_store.notifications.sync_to_async', side_effect=[mock_get_workspace])
     mock_threading = mocker.patch('tethysapp.app_store.notifications.threading')
     consumer = notificationsConsumer
+    consumer._authorized = True
     consumer.channel_layer_alias = "testlayer"
     channel_layers_setting = {
         "testlayer": {"BACKEND": "channels.layers.InMemoryChannelLayer"}
     }
     with override_settings(CHANNEL_LAYERS=channel_layers_setting):
-        consumer = notificationsConsumer.as_asgi()
-        communicator = WebsocketCommunicator(notificationsConsumer.as_asgi(), "GET", "install/notifications")
-        communicator.scope["user"] = mock_user
+        communicator = WebsocketCommunicator(consumer.as_asgi(), "GET", "install/notifications")
         connected, _ = await communicator.connect()
         assert connected
 
@@ -133,19 +123,15 @@ async def test_notificationsConsumer_receive_begin_install(mocker, caplog):
 
 
 @pytest.mark.asyncio
-async def test_notificationsConsumer_receive_invalid_type(mocker, caplog):
-    mock_duser = mocker.patch('tethysapp.app_store.notifications.User')
-    mock_duser.objects.get().has_perm.return_value = True
-    mock_user = MagicMock(id=1)
+async def test_notificationsConsumer_receive_invalid_type(caplog):
     consumer = notificationsConsumer
+    consumer._authorized = True
     consumer.channel_layer_alias = "testlayer"
     channel_layers_setting = {
         "testlayer": {"BACKEND": "channels.layers.InMemoryChannelLayer"}
     }
     with override_settings(CHANNEL_LAYERS=channel_layers_setting):
-        consumer = notificationsConsumer.as_asgi()
-        communicator = WebsocketCommunicator(notificationsConsumer.as_asgi(), "GET", "install/notifications")
-        communicator.scope["user"] = mock_user
+        communicator = WebsocketCommunicator(consumer.as_asgi(), "GET", "install/notifications")
         connected, _ = await communicator.connect()
         assert connected
